rename outcome.test.ts to encode-guarantee.test.ts
the other test cases belong in the exit-format repo
